#678 Refactored implementation of the default companion delete operations into a separate class.
